refactor: remove @scalar/swagger-parser, move code to @scalar/api-reference
