+++ .changeset/gold-buckets-heal.md
@@ -2,7 +2,6 @@
 "@scalar/fastify-api-reference": patch
 "@scalar/api-client-proxy": patch
 "@scalar/swagger-editor": patch
-"@scalar/swagger-parser": patch
 "@scalar/api-reference": patch
 "@scalar/echo-server": patch
 "@scalar/api-client": patch
