+++ packages/api-reference/package.json
@@ -58,9 +58,9 @@
     "@headlessui/vue": "^1.7.16",
     "@scalar/api-client": "workspace:*",
     "@scalar/components": "workspace:*",
+    "@scalar/openapi-parser": "^0.2.0",
     "@scalar/snippetz": "^0.1.5",
     "@scalar/swagger-editor": "workspace:*",
-    "@scalar/swagger-parser": "workspace:*",
     "@scalar/themes": "workspace:*",
     "@scalar/use-modal": "workspace:*",
     "@scalar/use-tooltip": "workspace:*",
@@ -104,6 +104,7 @@
     "@vitejs/plugin-vue": "^5.0.4",
     "@vitest/coverage-v8": "^1.2.2",
     "eslint": "^8.56.0",
+    "glob": "^10.3.5",
     "http-server": "^14.1.1",
     "postcss": "^8.4.31",
     "react": "^18.2.0",
