+++ packages/api-reference/src/helpers/index.ts
@@ -20,6 +20,7 @@ export * from './mapFromObject'
 export * from './mergeAllObjects'
 export * from './objectMerge'
 export * from './openClientFor'
+export * from './parse'
 export * from './prettyPrintJson'
 export * from './replaceVariables'
 export * from './scrollToId'
