+++ packages/api-reference/tests/fixtures/examples/CoinMarketCap.yaml
