+++ packages/api-reference/tests/fixtures/examples/Coinbase API.yaml
