+++ packages/api-reference/tests/fixtures/examples/DocuSign eSignature REST API.yaml
