+++ packages/api-reference/tests/fixtures/examples/Stripe API.yaml
