+++ packages/api-reference/tests/fixtures/examples/Zoom API.yaml
