+++ packages/api-reference/tests/fixtures/shopware.json
