+++ packages/api-reference/tests/fixtures/swagger.yaml
