--- packages/swagger-parser/CHANGELOG.md
@@ -1,248 +1,0 @@
-# @scalar/swagger-parser
-
-## 0.5.18
-
-### Patch Changes
-
-- 1669b27: chore: update dependencies
-
-## 0.5.17
-
-### Patch Changes
-
-- c6b95ca: chore: upgrade from Vite 4 to Vite 5
-
-## 0.5.16
-
-### Patch Changes
-
-- 92a8b9f: chore: update dependencies
-
-## 0.5.15
-
-### Patch Changes
-
-- 2c7e0b5c: feat: show webhooks
-
-## 0.5.14
-
-### Patch Changes
-
-- 81543f42: chore: add support for Node 18
-
-## 0.5.13
-
-### Patch Changes
-
-- a1319e8b: Switched internal component library to use ScalarComponents instead
-
-## 0.5.12
-
-### Patch Changes
-
-- 35dffe4: chore: include @scalar dependencies in the bundle
-- 5c66be8: chore: replace imports pointing to src with relative paths
-
-## 0.5.11
-
-### Patch Changes
-
-- 55246e5: chore: rewrite all package.jsons (including their prod, dev and peer dependency requirements)
-
-## 0.5.10
-
-### Patch Changes
-
-- b8dc5a5: feat: add pathParameters to the result
-
-## 0.5.9
-
-### Patch Changes
-
-- a1575f2: chore: don’t use the operationId as a name fallback
-
-## 0.5.8
-
-### Patch Changes
-
-- 3d6027c: chore: use the summary as the title, fall back to operationId, and then the path
-
-## 0.5.7
-
-### Patch Changes
-
-- 54098d7: chore: better fallbacks for the operation.name
-
-## 0.5.6
-
-### Patch Changes
-
-- dab853a: chore: remove unused dependencies
-
-## 0.5.5
-
-### Patch Changes
-
-- db7f5f57: feat: add analyze helper to get stats for Swagger/OpenAPI specs
-- d87566fd: refactor: new section components, new sidebar
-- 8be53df2: fix: avoid timeouts with large files, downgrade @apidevtools/json-schema-ref-parser
-- 4eaee866: feat: add preflight helper function to quickly analyze specs without parsing them
-- 06b46553: refactor: rename parseSwaggerFile to just parse
-
-## 0.5.4
-
-### Patch Changes
-
-- fd6cf1bb: fix: too much recursion
-
-## 0.5.3
-
-### Patch Changes
-
-- af4ec8d3: update discord links
-
-## 0.5.2
-
-### Patch Changes
-
-- 33604a1b: chore: fix types
-- 1cbcabd7: lots of amazing fixes from when we missed last patch
-- d7b12610: fix: handle specs without example responses
-
-## 0.5.1
-
-### Patch Changes
-
-- 048555c6: feat: add CHANGELOG.md to the package
-
-## 0.5.0
-
-### Minor Changes
-
-- 87fda5af: chore: update dependencies
-
-### Patch Changes
-
-- ef080a26: chore: format package.json
-
-## 0.4.3
-
-### Patch Changes
-
-- d51aebe5: chore: remove source files from packages
-
-## 0.4.2
-
-### Patch Changes
-
-- 4165b251: feat: improve rendering of complex swagger files
-
-## 0.4.1
-
-### Patch Changes
-
-- 10498d9d: chore: add README
-- 494083a0: add more information to the package.json
-
-## 0.4.0
-
-### Minor Changes
-
-- 5057e213: make npm packages public :-)
-
-## 0.3.0
-
-### Minor Changes
-
-- b0f5221: manually releasing all packages to make sure the lastest version is on npm
-
-## 0.2.0
-
-### Minor Changes
-
-- a ton of fixes, new fastify plugin
-
-## 0.1.14
-
-### Patch Changes
-
-- fix: Node polyfill issues (process undefined)
-
-## 0.1.13
-
-### Patch Changes
-
-- add offline search and ui improvements
-- add offline search, improve ui
-
-## 0.1.12
-
-### Patch Changes
-
-- feat: add a standalone version
-
-## 0.1.11
-
-### Patch Changes
-
-- fix sidebar height issue
-
-## 0.1.10
-
-### Patch Changes
-
-- fix: add proper type check on window
-
-## 0.1.9
-
-### Patch Changes
-
-- fix: add window polyfill wrapper
-
-## 0.1.8
-
-### Patch Changes
-
-- add fallback for operation id
-
-## 0.1.7
-
-### Patch Changes
-
-- css variable fixes and hocuspocusurl prop
-
-## 0.1.6
-
-### Patch Changes
-
-- fix: issues with empty responses, minor styling issues
-
-## 0.1.5
-
-### Patch Changes
-
-- style: improve default style
-
-## 0.1.4
-
-### Patch Changes
-
-- fix: externalize dependencies
-
-## 0.1.3
-
-### Patch Changes
-
-- fix: add CSS file to swagger editor
-
-## 0.1.2
-
-### Patch Changes
-
-- 7cd41bc: fix: point main entry to dist folder
-
-## 0.1.1
-
-### Patch Changes
-
-- f46e320: Initial release 👀
