--- packages/swagger-parser/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@scalar/swagger-parser",
-  "description": "parse OpenAPI specs",
-  "license": "MIT",
-  "author": "Scalar (https://github.com/scalar)",
-  "homepage": "https://github.com/scalar/scalar",
-  "bugs": "https://github.com/scalar/scalar/issues/new/choose",
-  "keywords": [
-    "openapi",
-    "parser",
-    "swagger"
-  ],
-  "version": "0.5.18",
-  "engines": {
-    "node": ">=18"
-  },
-  "scripts": {
-    "build": "vite build && pnpm types:build && tsc-alias -p tsconfig.build.json",
-    "dev": "vite",
-    "lint:check": "eslint .",
-    "lint:fix": "eslint .  --fix",
-    "test": "vitest",
-    "types:build": "tsc -p tsconfig.build.json",
-    "types:check": "tsc --noEmit --skipLibCheck"
-  },
-  "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    "import": "./dist/index.js"
-  },
-  "files": [
-    "dist",
-    "CHANGELOG.md"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/scalar/scalar.git",
-    "directory": "packages/swagger-parser"
-  },
-  "dependencies": {
-    "@scalar/openapi-parser": "^0.2.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.6.3",
-    "@vitejs/plugin-vue": "^5.0.4",
-    "@vitest/coverage-v8": "^1.2.2",
-    "glob": "^10.3.5",
-    "openapi-types": "^12.1.3",
-    "tsc-alias": "^1.8.8",
-    "typescript": "^5.2.2",
-    "vite": "^5.1.1",
-    "vitest": "^1.2.2"
-  }
-}
