--- packages/swagger-parser/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "target": "ES2022" /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */,
-    "module": "ESNext" /* Specify what module code is generated. */,
-    "moduleResolution": "Node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "esModuleInterop": true /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */,
-    "forceConsistentCasingInFileNames": true /* Ensure that casing is correct in imports. */,
-    "strict": true /* Enable all strict type-checking options. */,
-    "strictNullChecks": true,
-    "types": ["vite/client"]
-  },
-  "exclude": ["dist", "vite.config.ts", "node_modules"],
-  "ts-node": {
-    "require": ["tsconfig-paths/register"]
-  }
-}
