--- packages/swagger-parser/vite.config.ts
@@ -1,45 +1,0 @@
-import path from 'path'
-import { defineConfig } from 'vitest/config'
-
-import pkg from './package.json'
-
-export default defineConfig({
-  build: {
-    lib: {
-      entry: 'src/index.ts',
-      name: '@scalar/swagger-parser',
-      fileName: 'index',
-      formats: ['es'],
-    },
-    rollupOptions: {
-      /**
-       * Make sure to also externalize any dependencies that you do not want to bundle into your library
-       */
-      external: [
-        'vue',
-        ...Object.keys(pkg.dependencies || {}).filter(
-          (item) => !item.startsWith('@scalar'),
-        ),
-      ],
-    },
-  },
-  resolve: {
-    alias: [
-      // Resolve the uncompiled source code for all @scalar packages
-      // It’s working with the alias, too. It’s just required to enable HMR.
-      {
-        // Resolve the uncompiled source code for all @scalar packages
-        // @scalar/* -> packages/*/
-        // (not @scalar/*/style.css)
-        find: /^@scalar\/(?!(openapi-parser|snippetz|components\/style\.css|components\b))(.+)/,
-        replacement: path.resolve(__dirname, '../$2/src/index.ts'),
-      },
-    ],
-  },
-  test: {
-    coverage: {
-      enabled: true,
-      reporter: 'text',
-    },
-  },
-})
